Move block-related commands to eth_block.go
